Move ray solver functions to new file
